requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
python-socketio==5.14.3
python-engineio==4.12.3
pyOpenSSL==25.3.0
eventlet==0.36.1
